Move start_config out of build_commands file
