+++ freqtrade/commands/__init__.py
@@ -7,9 +7,9 @@ Note: Be careful with file-scoped imports in these subfiles.
     as they are parsed on startup, nothing containing optional modules should be loaded.
 """
 from freqtrade.commands.arguments import Arguments
+from freqtrade.commands.build_config_commands import start_new_config
 from freqtrade.commands.data_commands import start_download_data
 from freqtrade.commands.deploy_commands import (start_create_userdir,
-                                                start_new_config,
                                                 start_new_hyperopt,
                                                 start_new_strategy)
 from freqtrade.commands.hyperopt_commands import (start_hyperopt_list,
