+++ freqtrade/templates/subtemplates/exchange_generic.j2
@@ -3,7 +3,7 @@
 "secret": "{{ exchange_secret }}",
 "ccxt_config": {"enableRateLimit": true},
 "ccxt_async_config": {
-    "enableRateLimit": true,
+    "enableRateLimit": true
 },
 "pair_whitelist": [
